small tidy up of helper functions/factories in TCC schema
